Add a manifest
The manifest is a list of all the input files. This allows the server to
easily get an up to date list of what is available, and to refer to the
files by name rather than by full URL, to make sure the compute images
get the files from a trusted source.